refactor(agent): move evaluation isolation out of product
Keep normal product Web and shell behavior independent from benchmark execution policy. Network egress controls, repository-history sanitization, and benchmark leakage auditing belong in the evaluation pipeline rather than shared agent tools.

Changes:

- restore WebFetch and WebSearch access to GitHub, Sourcegraph, redirects, and code-hosting search results

- remove shell domain, Git network, remote-ref, commit-hash, and GitHub CLI restrictions while retaining SQLite WAL safety and user-authored edit constraints

- remove the eval-oriented AI request lifecycle wrapper while preserving the existing effective-options audit and provider request fixes

- split edit-constraint state, shell target parsing, and tests into focused modules without changing the public facade

- replace the stale benchmark retrospective with the current product boundary design and remove the evaluation-network research note

Verification:

- cargo check --workspace

- cargo check -p bitfun-ai-adapters

- cargo test -p bitfun-core edit_constraint_guard --lib (22 passed)

- cargo test -p bitfun-core shell_command_safety --lib (4 passed)

- focused WebFetch, WebSearch, and normal Git/shell allow tests passed

The existing bitfun-ai-adapters lib-test AIClient::new Result migration issue is outside this change; non-test adapter builds pass.
