+++ src/crates/assembly/core/src/agentic/execution/edit_constraint_guard.rs
@@ -6,11 +6,9 @@
 //! file mutation is appended to a session-scoped JSONL telemetry stream.
 
 use log::warn;
-use regex::Regex;
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use serde_json::{json, Value};
 use sha2::{Digest, Sha256};
-use std::collections::HashSet;
 use std::fs::{self, OpenOptions};
 use std::io::Write;
 use std::path::Path;
@@ -23,6 +21,16 @@ use crate::agentic::tools::framework::{ToolUseContext, ValidationResult};
 use crate::infrastructure::ai::get_global_ai_client_factory;
 use crate::util::json_extract::extract_json_from_ai_response;
 use crate::util::types::Message;
+
+mod model;
+mod shell_targets;
+
+pub use model::{
+    AgentCreatedPathRecord, ConstraintExtractionRecord, ConstraintMatcher,
+    ConstraintOperationScope, ConstraintRevocation, ConstraintSource, EditConstraintState,
+    ExtractedConstraint, ExtractionFailure, ExtractionStatus, ModelExtractionStatus,
+};
+use shell_targets::explicit_bash_mutation_targets;
 
 pub const EDIT_CONSTRAINT_METADATA_KEY: &str = "editConstraintGuard";
 const EDIT_CONSTRAINT_SCHEMA_VERSION: u32 = 5;
@@ -73,348 +81,6 @@ Respond with ONLY a fenced ```json code block containing this exact shape:
 ```
 If the latest message changes nothing, return empty `additions` and
 `revocations`."#;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum ConstraintSource {
-    Deterministic,
-    Model,
-    #[default]
-    Legacy,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExtractedConstraint {
-    #[serde(default)]
-    pub id: String,
-    /// Human-readable paraphrase shown to the agent in the rejection message.
-    pub description: String,
-    #[serde(default)]
-    pub operation_scope: ConstraintOperationScope,
-    pub matcher: ConstraintMatcher,
-    #[serde(default)]
-    pub source: ConstraintSource,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source_text: Option<String>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum ConstraintOperationScope {
-    #[default]
-    All,
-    DeleteOnly,
-}
-
-impl ConstraintOperationScope {
-    fn applies_to(self, operation: &str) -> bool {
-        match self {
-            Self::All => true,
-            Self::DeleteOnly => matches!(operation, "delete" | "recursive_delete"),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum ConstraintMatcher {
-    TestFiles,
-    PathContains {
-        substrings: Vec<String>,
-    },
-    PathUnderDir {
-        dirs: Vec<String>,
-    },
-    Extension {
-        exts: Vec<String>,
-    },
-    /// Recorded for analysis but never enforced.
-    Unmatched,
-}
-
-impl ConstraintMatcher {
-    pub fn matches(&self, file_path: &str) -> bool {
-        let normalized = file_path.replace('\\', "/");
-        match self {
-            ConstraintMatcher::TestFiles => is_test_file(&normalized),
-            ConstraintMatcher::PathContains { substrings } => substrings
-                .iter()
-                .any(|value| !value.is_empty() && normalized.contains(value.as_str())),
-            ConstraintMatcher::PathUnderDir { dirs } => dirs.iter().any(|dir| {
-                let dir = dir.trim_matches('/');
-                !dir.is_empty()
-                    && (normalized == dir
-                        || normalized.starts_with(&format!("{dir}/"))
-                        || normalized.contains(&format!("/{dir}/")))
-            }),
-            ConstraintMatcher::Extension { exts } => exts
-                .iter()
-                .any(|extension| !extension.is_empty() && normalized.ends_with(extension)),
-            ConstraintMatcher::Unmatched => false,
-        }
-    }
-
-    fn enforceable(&self) -> bool {
-        !matches!(self, ConstraintMatcher::Unmatched)
-    }
-}
-
-fn is_test_file(path: &str) -> bool {
-    let normalized = path.replace('\\', "/");
-    let lowercase = normalized.to_lowercase();
-    let name = Path::new(&lowercase)
-        .file_name()
-        .and_then(|name| name.to_str())
-        .unwrap_or_default();
-    let stem = Path::new(name)
-        .file_stem()
-        .and_then(|value| value.to_str())
-        .unwrap_or_default();
-
-    stem.starts_with("test_")
-        || stem.starts_with("test-")
-        || stem.ends_with("_test")
-        || stem.ends_with("-test")
-        || stem.ends_with("_tests")
-        || stem.ends_with("_spec")
-        || stem.ends_with("-spec")
-        || name.contains(".test.")
-        || name.contains(".spec.")
-        || lowercase
-            .split('/')
-            .any(|segment| matches!(segment, "tests" | "test" | "__tests__" | "spec" | "specs"))
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ExtractionStatus {
-    Extracted,
-    NoConstraints,
-    Failed,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum ModelExtractionStatus {
-    #[default]
-    NotRun,
-    Parsed,
-    Failed,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ConstraintRevocation {
-    pub constraint_id: String,
-    pub description: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExtractionFailure {
-    pub stage: String,
-    pub reason: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ConstraintExtractionRecord {
-    pub message_sha256: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub dialog_turn_id: Option<String>,
-    pub status: ExtractionStatus,
-    pub constraints: Vec<ExtractedConstraint>,
-    pub deterministic_constraint_count: usize,
-    pub model_attempts: usize,
-    /// Snapshot of the active ids shown to fast for this extraction.
-    #[serde(default)]
-    pub active_constraint_ids: Vec<String>,
-    /// Whether this turn originated at a real user submission and can therefore
-    /// relax an existing user-authored edit constraint.
-    #[serde(default)]
-    pub revocation_authorized: bool,
-    #[serde(default)]
-    pub model_status: ModelExtractionStatus,
-    /// Exact additions parsed from the fast model, before deterministic
-    /// additions are merged in.
-    #[serde(default)]
-    pub model_constraints: Vec<ExtractedConstraint>,
-    /// Exact revocation requests parsed from the fast model. Invalid ids remain
-    /// here for telemetry but are never applied.
-    #[serde(default)]
-    pub model_revocations: Vec<ConstraintRevocation>,
-    /// Revocations validated against the active constraint ids supplied to the
-    /// model. Only these ids are applied to session state.
-    #[serde(default)]
-    pub revoked_constraint_ids: Vec<String>,
-    #[serde(default)]
-    pub unmatched_revocation_ids: Vec<String>,
-    pub input_chars: usize,
-    pub prompt_chars: usize,
-    pub input_truncated: bool,
-    pub latency_ms: u64,
-    pub extracted_at_ms: u64,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub failure: Option<ExtractionFailure>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub response_excerpt: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct EditConstraintState {
-    pub schema_version: u32,
-    #[serde(default)]
-    pub constraints: Vec<ExtractedConstraint>,
-    #[serde(default)]
-    pub extractions: Vec<ConstraintExtractionRecord>,
-    /// Paths first created through direct agent file tools in this session.
-    /// They remain distinct from repository files across session restoration.
-    #[serde(default)]
-    pub agent_created_paths: Vec<String>,
-    /// Turn-scoped provenance used to rewind helper-file permissions when a
-    /// session is rolled back. `agent_created_paths` remains for backwards
-    /// compatibility and is rebuilt from these records when possible.
-    #[serde(default)]
-    pub agent_created_path_records: Vec<AgentCreatedPathRecord>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct AgentCreatedPathRecord {
-    pub path: String,
-    pub dialog_turn_id: String,
-}
-
-impl Default for EditConstraintState {
-    fn default() -> Self {
-        Self {
-            schema_version: EDIT_CONSTRAINT_SCHEMA_VERSION,
-            constraints: Vec::new(),
-            extractions: Vec::new(),
-            agent_created_paths: Vec::new(),
-            agent_created_path_records: Vec::new(),
-        }
-    }
-}
-
-impl EditConstraintState {
-    pub fn merge_extraction(&mut self, extraction: ConstraintExtractionRecord) {
-        self.schema_version = EDIT_CONSTRAINT_SCHEMA_VERSION;
-        self.constraints.retain(|constraint| {
-            !extraction
-                .revoked_constraint_ids
-                .iter()
-                .any(|constraint_id| constraint_id == &constraint.id)
-        });
-        for constraint in &extraction.constraints {
-            if !self.constraints.iter().any(|existing| {
-                existing.matcher == constraint.matcher
-                    && existing.operation_scope == constraint.operation_scope
-            }) {
-                self.constraints.push(constraint.clone());
-            }
-        }
-        self.extractions.push(extraction);
-    }
-
-    pub fn message_processed(&self, dialog_turn_id: &str, message_sha256: &str) -> bool {
-        self.extractions.iter().any(|record| {
-            record.dialog_turn_id.as_deref() == Some(dialog_turn_id)
-                && record.message_sha256 == message_sha256
-                && record.status != ExtractionStatus::Failed
-        })
-    }
-
-    pub fn latest_status(&self) -> Option<ExtractionStatus> {
-        self.extractions.last().map(|record| record.status)
-    }
-
-    pub fn has_enforceable_constraints(&self) -> bool {
-        self.constraints
-            .iter()
-            .any(|constraint| constraint.matcher.enforceable())
-    }
-
-    pub fn remember_agent_created_paths(
-        &mut self,
-        paths: impl IntoIterator<Item = String>,
-        dialog_turn_id: &str,
-    ) {
-        for path in paths {
-            let normalized = path.replace('\\', "/");
-            if !normalized.is_empty() && !self.agent_created_paths.contains(&normalized) {
-                self.agent_created_paths.push(normalized.clone());
-            }
-            if !normalized.is_empty()
-                && !dialog_turn_id.is_empty()
-                && !self.agent_created_path_records.iter().any(|record| {
-                    record.path == normalized && record.dialog_turn_id == dialog_turn_id
-                })
-            {
-                self.agent_created_path_records
-                    .push(AgentCreatedPathRecord {
-                        path: normalized,
-                        dialog_turn_id: dialog_turn_id.to_string(),
-                    });
-            }
-        }
-    }
-
-    pub fn forget_agent_created_paths_under(&mut self, paths: &[String]) {
-        self.agent_created_paths.retain(|created| {
-            !paths.iter().any(|path| {
-                let path = path.trim_end_matches('/');
-                created == path || created.starts_with(&format!("{path}/"))
-            })
-        });
-        self.agent_created_path_records.retain(|record| {
-            !paths.iter().any(|path| {
-                let path = path.trim_end_matches('/');
-                record.path == path || record.path.starts_with(&format!("{path}/"))
-            })
-        });
-    }
-
-    fn is_agent_created_path(&self, paths: &[String]) -> bool {
-        paths
-            .iter()
-            .any(|path| self.agent_created_paths.contains(path))
-    }
-
-    /// Rebuilds the state from events belonging to turns that survive a
-    /// session rollback. Older provenance entries did not carry a turn id, so
-    /// they are deliberately discarded rather than granting a stale exemption
-    /// to a file created only in a rolled-back future turn.
-    pub fn rollback_to_surviving_turns(&mut self, surviving_turn_ids: &HashSet<String>) {
-        let retained_extractions = self
-            .extractions
-            .iter()
-            .filter(|record| {
-                record
-                    .dialog_turn_id
-                    .as_ref()
-                    .is_some_and(|turn_id| surviving_turn_ids.contains(turn_id))
-            })
-            .cloned()
-            .collect::<Vec<_>>();
-        let retained_path_records = self
-            .agent_created_path_records
-            .iter()
-            .filter(|record| surviving_turn_ids.contains(&record.dialog_turn_id))
-            .cloned()
-            .collect::<Vec<_>>();
-
-        self.schema_version = EDIT_CONSTRAINT_SCHEMA_VERSION;
-        self.constraints.clear();
-        self.extractions.clear();
-        for extraction in retained_extractions {
-            self.merge_extraction(extraction);
-        }
-        self.agent_created_paths.clear();
-        self.agent_created_path_records = retained_path_records;
-        for record in &self.agent_created_path_records {
-            if !self.agent_created_paths.contains(&record.path) {
-                self.agent_created_paths.push(record.path.clone());
-            }
-        }
-    }
-}
 
 #[derive(Debug, Deserialize)]
 #[serde(deny_unknown_fields)]
@@ -1360,220 +1026,6 @@ pub fn check_bash_command(context: &ToolUseContext, command: &str) -> Option<Val
     None
 }
 
-fn explicit_bash_mutation_targets(command: &str) -> Vec<String> {
-    let mut targets = Vec::new();
-    // Python `-c` programs commonly contain semicolons inside the quoted
-    // program. Scan the complete command before shell-level segmentation so a
-    // later `Path(...).write_text()` expression keeps its Python context.
-    push_python_mutation_targets(&mut targets, command);
-    push_node_mutation_targets(&mut targets, command);
-    for segment in command
-        .split(['\n', ';', '|'])
-        .flat_map(|part| part.split("&&"))
-        .flat_map(|part| part.split("||"))
-    {
-        let words = segment
-            .split_whitespace()
-            .map(|word| {
-                word.trim_matches(|c: char| matches!(c, '\'' | '"' | '(' | ')' | '[' | ']'))
-            })
-            .filter(|word| !word.is_empty())
-            .collect::<Vec<_>>();
-        if words.is_empty() {
-            continue;
-        }
-
-        for (index, word) in words.iter().enumerate() {
-            let redirection = word.trim_start_matches(|c| matches!(c, '0'..='9'));
-            if matches!(redirection, ">" | ">>" | "1>" | "1>>") {
-                if let Some(path) = words.get(index + 1) {
-                    push_bash_target(&mut targets, path);
-                }
-            } else if let Some(path) = redirection
-                .strip_prefix(">>")
-                .or_else(|| redirection.strip_prefix('>'))
-            {
-                if !path.is_empty() {
-                    push_bash_target(&mut targets, path);
-                }
-            }
-        }
-
-        let Some(command_index) = words.iter().position(|word| !word.contains('=')) else {
-            continue;
-        };
-        let command_name = Path::new(words[command_index])
-            .file_name()
-            .and_then(|name| name.to_str())
-            .unwrap_or(words[command_index])
-            .to_ascii_lowercase();
-        let arguments = &words[command_index + 1..];
-        match command_name.as_str() {
-            "tee" => {
-                for argument in arguments
-                    .iter()
-                    .filter(|argument| !argument.starts_with('-'))
-                {
-                    push_bash_target(&mut targets, argument);
-                }
-            }
-            "cp" | "install" => {
-                if let Some(path) = arguments
-                    .iter()
-                    .rev()
-                    .find(|argument| !argument.starts_with('-'))
-                {
-                    push_bash_target(&mut targets, path);
-                }
-            }
-            "mv" => {
-                // Moving a protected source removes it from its original
-                // location, so both sides are mutation targets. The previous
-                // destination-only handling let `mv tests/a.rs src/a.rs`
-                // bypass a test-file constraint.
-                for argument in arguments
-                    .iter()
-                    .filter(|argument| !argument.starts_with('-'))
-                {
-                    push_bash_target(&mut targets, argument);
-                }
-            }
-            "touch" | "truncate" | "rm" | "rmdir" | "unlink" => {
-                for argument in arguments
-                    .iter()
-                    .filter(|argument| !argument.starts_with('-'))
-                {
-                    push_bash_target(&mut targets, argument);
-                }
-            }
-            "sed" | "perl" => {
-                if arguments
-                    .iter()
-                    .any(|argument| *argument == "-i" || argument.starts_with("-i"))
-                {
-                    let mut script_seen = false;
-                    for argument in arguments
-                        .iter()
-                        .filter(|argument| !argument.starts_with('-'))
-                    {
-                        if !script_seen {
-                            script_seen = true;
-                            continue;
-                        }
-                        if argument.starts_with('/')
-                            || argument.starts_with("./")
-                            || argument.starts_with("../")
-                            || argument.contains('.')
-                            || argument.starts_with("test/")
-                            || argument.starts_with("tests/")
-                        {
-                            push_bash_target(&mut targets, argument);
-                        }
-                    }
-                }
-            }
-            "git" => push_git_mutation_targets(&mut targets, arguments),
-            _ => {}
-        }
-    }
-    targets
-}
-
-fn push_git_mutation_targets(targets: &mut Vec<String>, arguments: &[&str]) {
-    let Some((subcommand_index, subcommand)) = arguments
-        .iter()
-        .enumerate()
-        .find(|(_, argument)| !argument.starts_with('-'))
-    else {
-        return;
-    };
-    let remaining = &arguments[subcommand_index + 1..];
-    match *subcommand {
-        "mv" | "rm" | "restore" => {
-            for argument in remaining
-                .iter()
-                .filter(|argument| !argument.starts_with('-'))
-            {
-                push_bash_target(targets, argument);
-            }
-        }
-        "checkout" => {
-            // `git checkout <ref>` is not a path mutation by itself. The
-            // pathspec form is unambiguous only after `--`.
-            if let Some(separator) = remaining.iter().position(|argument| *argument == "--") {
-                for argument in remaining[separator + 1..]
-                    .iter()
-                    .filter(|argument| !argument.starts_with('-'))
-                {
-                    push_bash_target(targets, argument);
-                }
-            }
-        }
-        _ => {}
-    }
-}
-
-fn push_python_mutation_targets(targets: &mut Vec<String>, segment: &str) {
-    static OPEN_FOR_WRITE: OnceLock<Regex> = OnceLock::new();
-    static PATH_MUTATION: OnceLock<Regex> = OnceLock::new();
-    let open_for_write = OPEN_FOR_WRITE.get_or_init(|| {
-        Regex::new(r#"(?i)\bopen\s*\(\s*["']([^"']+)["']\s*,\s*["'][wax][^"']*["']"#)
-            .expect("valid Python open-for-write regex")
-    });
-    let path_mutation = PATH_MUTATION.get_or_init(|| {
-        Regex::new(
-            r#"(?i)\bPath\s*\(\s*["']([^"']+)["']\s*\)\s*\.\s*(?:write_text|write_bytes|unlink|rename|replace)\s*\("#,
-        )
-        .expect("valid pathlib mutation regex")
-    });
-
-    for captures in open_for_write
-        .captures_iter(segment)
-        .chain(path_mutation.captures_iter(segment))
-    {
-        if let Some(path) = captures.get(1) {
-            push_bash_target(targets, path.as_str());
-        }
-    }
-}
-
-fn push_node_mutation_targets(targets: &mut Vec<String>, segment: &str) {
-    static SINGLE_PATH_MUTATION: OnceLock<Regex> = OnceLock::new();
-    static TWO_PATH_MUTATION: OnceLock<Regex> = OnceLock::new();
-    let single_path_mutation = SINGLE_PATH_MUTATION.get_or_init(|| {
-        Regex::new(
-            r#"(?i)\b(?:fs\s*\.\s*)?(?:writefilesync|appendfilesync|unlinksync|rmsync)\s*\(\s*["']([^"']+)["']"#,
-        )
-        .expect("valid Node single-path mutation regex")
-    });
-    let two_path_mutation = TWO_PATH_MUTATION.get_or_init(|| {
-        Regex::new(
-            r#"(?i)\b(?:fs\s*\.\s*)?(?:renamesync|copyfilesync)\s*\(\s*["']([^"']+)["']\s*,\s*["']([^"']+)["']"#,
-        )
-        .expect("valid Node two-path mutation regex")
-    });
-
-    for captures in single_path_mutation.captures_iter(segment) {
-        if let Some(path) = captures.get(1) {
-            push_bash_target(targets, path.as_str());
-        }
-    }
-    for captures in two_path_mutation.captures_iter(segment) {
-        for index in [1, 2] {
-            if let Some(path) = captures.get(index) {
-                push_bash_target(targets, path.as_str());
-            }
-        }
-    }
-}
-
-fn push_bash_target(targets: &mut Vec<String>, raw_path: &str) {
-    let path = raw_path.trim_matches(|c: char| matches!(c, '\'' | '"' | ','));
-    if !path.is_empty() && !targets.iter().any(|existing| existing == path) {
-        targets.push(path.to_string());
-    }
-}
-
 /// Checks the target and every non-symlink descendant before recursive delete.
 /// Inspection failures are fail-closed only when an enforceable constraint is
 /// active, because otherwise there is no protected path to discover.
@@ -1958,597 +1410,4 @@ pub fn record_mutation_applied(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::agentic::tools::ToolRuntimeRestrictions;
-    use crate::agentic::WorkspaceBinding;
-    use std::collections::HashMap;
-
-    fn constraint(description: &str, matcher: ConstraintMatcher) -> ExtractedConstraint {
-        ExtractedConstraint {
-            id: format!("test:{description}"),
-            description: description.to_string(),
-            operation_scope: ConstraintOperationScope::All,
-            matcher,
-            source: ConstraintSource::Legacy,
-            source_text: None,
-        }
-    }
-
-    #[test]
-    fn test_files_matcher_covers_common_conventions() {
-        let matcher = ConstraintMatcher::TestFiles;
-        for path in [
-            "report/util_test.go",
-            "pkg/foo/test_bar.py",
-            "pkg/foo/bar_test.py",
-            "src/foo.test.tsx",
-            "src/foo.spec.ts",
-            "spec/models/user_spec.rb",
-            "pkg/foo_test.cc",
-            "src/foo-test.js",
-            "src/test-widget.ts",
-            "test/components/Foo-test.tsx",
-            "__tests__/foo.js",
-            "TEST/UPPER.spec.ts",
-        ] {
-            assert!(matcher.matches(path), "expected test path: {path}");
-        }
-        assert!(!matcher.matches("src/foo.ts"));
-        assert!(!matcher.matches("report/util.go"));
-    }
-
-    #[test]
-    fn deterministic_extractor_recognizes_swebench_wording() {
-        let message = "I've already taken care of all changes to any of the test files. You DON'T have to modify the testing logic or any of the tests. Keep changes minimal and limited to non-tests.";
-        let extracted = deterministic_test_constraint(message).expect("test constraint");
-        assert_eq!(extracted.matcher, ConstraintMatcher::TestFiles);
-        assert_eq!(extracted.source, ConstraintSource::Deterministic);
-        assert!(extracted.source_text.is_some());
-    }
-
-    #[test]
-    fn deterministic_extractor_does_not_confuse_do_not_run_tests() {
-        assert!(deterministic_test_constraint("Do not run the tests on Windows.").is_none());
-    }
-
-    #[test]
-    fn deterministic_extractor_recognizes_unchanged_and_non_test_only_wording() {
-        for message in [
-            "Keep test files unchanged.",
-            "Tests must remain unchanged.",
-            "Only modify non-test files.",
-            "测试文件保持不变。",
-        ] {
-            assert!(
-                deterministic_test_constraint(message).is_some(),
-                "expected deterministic constraint for: {message}"
-            );
-        }
-    }
-
-    #[test]
-    fn deterministic_extractor_does_not_turn_explicit_relaxation_into_a_prohibition() {
-        for message in [
-            "You can modify tests now.",
-            "Test files are allowed to be modified.",
-            "现在可以修改测试文件。",
-        ] {
-            assert!(
-                deterministic_test_constraint(message).is_none(),
-                "expected no deterministic prohibition for: {message}"
-            );
-        }
-    }
-
-    #[test]
-    fn long_prompt_keeps_both_ends() {
-        let input = format!("start{}do not modify tests", "x".repeat(MAX_PROMPT_CHARS));
-        let (truncated, was_truncated) = truncate_for_extraction(&input);
-        assert!(was_truncated);
-        assert!(truncated.starts_with("start"));
-        assert!(truncated.ends_with("do not modify tests"));
-    }
-
-    #[test]
-    fn matchers_cover_paths_extensions_and_unmatched() {
-        assert!(ConstraintMatcher::PathContains {
-            substrings: vec!["package-lock.json".to_string()]
-        }
-        .matches("frontend/package-lock.json"));
-        assert!(ConstraintMatcher::PathUnderDir {
-            dirs: vec!["migrations".to_string()]
-        }
-        .matches("db/migrations/0002_add_column.sql"));
-        assert!(ConstraintMatcher::Extension {
-            exts: vec![".lock".to_string()]
-        }
-        .matches("Cargo.lock"));
-        assert!(!ConstraintMatcher::Unmatched.matches("anything.go"));
-    }
-
-    #[test]
-    fn terminal_preflight_finds_explicit_mutation_targets() {
-        assert_eq!(
-            explicit_bash_mutation_targets("sed -i 's/old/new/' tests/example.rs"),
-            vec!["tests/example.rs".to_string()]
-        );
-        assert_eq!(
-            explicit_bash_mutation_targets("printf x > test/unit/output.txt && touch src/lib.rs"),
-            vec!["test/unit/output.txt".to_string(), "src/lib.rs".to_string()]
-        );
-        assert_eq!(
-            explicit_bash_mutation_targets("cargo test -p core"),
-            Vec::<String>::new()
-        );
-        assert_eq!(
-            explicit_bash_mutation_targets(
-                r#"python3 -c \"open('/app/test/unit/example_test.py', 'w').write('x')\""#
-            ),
-            vec!["/app/test/unit/example_test.py".to_string()]
-        );
-        assert_eq!(
-            explicit_bash_mutation_targets(
-                r#"python -c \"from pathlib import Path; Path('tests/repro_test.py').write_text('x')\""#
-            ),
-            vec!["tests/repro_test.py".to_string()]
-        );
-        assert_eq!(
-            explicit_bash_mutation_targets("mv tests/existing_test.py src/existing.py"),
-            vec![
-                "tests/existing_test.py".to_string(),
-                "src/existing.py".to_string()
-            ]
-        );
-        assert_eq!(
-            explicit_bash_mutation_targets(
-                r#"node -e \"require('fs').writeFileSync('tests/example.test.js', 'x')\""#
-            ),
-            vec!["tests/example.test.js".to_string()]
-        );
-        assert_eq!(
-            explicit_bash_mutation_targets("git mv tests/example.rs src/example.rs"),
-            vec!["tests/example.rs".to_string(), "src/example.rs".to_string()]
-        );
-        assert_eq!(
-            explicit_bash_mutation_targets("git checkout HEAD -- tests/example.rs"),
-            vec!["tests/example.rs".to_string()]
-        );
-    }
-
-    #[test]
-    fn fast_response_parser_requires_the_observable_update_schema() {
-        let valid = r#"{
-            "additions": [{
-                "description": "do not modify tests",
-                "matcher": {"kind": "test_files"}
-            }],
-            "revocations": [{
-                "constraint_id": "deterministic:test_files",
-                "description": "tests may now be modified"
-            }]
-        }"#;
-        let parsed: ExtractionResponse = serde_json::from_str(valid).expect("valid schema");
-        assert_eq!(parsed.additions.len(), 1);
-        assert_eq!(parsed.revocations.len(), 1);
-
-        assert!(serde_json::from_str::<ExtractionResponse>(
-            r#"{"constraints": [], "revocations": []}"#
-        )
-        .is_err());
-        assert!(serde_json::from_str::<ExtractionResponse>(r#"{"additions": []}"#).is_err());
-    }
-
-    #[test]
-    fn state_distinguishes_failed_from_processed_extraction() {
-        let mut state = EditConstraintState::default();
-        let failed = ConstraintExtractionRecord {
-            message_sha256: "hash".to_string(),
-            dialog_turn_id: Some("turn-1".to_string()),
-            status: ExtractionStatus::Failed,
-            constraints: Vec::new(),
-            deterministic_constraint_count: 0,
-            model_attempts: 2,
-            active_constraint_ids: Vec::new(),
-            revocation_authorized: true,
-            model_status: ModelExtractionStatus::Failed,
-            model_constraints: Vec::new(),
-            model_revocations: Vec::new(),
-            revoked_constraint_ids: Vec::new(),
-            unmatched_revocation_ids: Vec::new(),
-            input_chars: 10,
-            prompt_chars: 10,
-            input_truncated: false,
-            latency_ms: 1,
-            extracted_at_ms: 1,
-            failure: Some(ExtractionFailure {
-                stage: "schema_validation".to_string(),
-                reason: "bad json".to_string(),
-            }),
-            response_excerpt: None,
-        };
-        state.merge_extraction(failed);
-        assert!(!state.message_processed("turn-1", "hash"));
-
-        let mut completed = state.extractions[0].clone();
-        completed.status = ExtractionStatus::NoConstraints;
-        completed.failure = None;
-        state.merge_extraction(completed);
-        assert!(state.message_processed("turn-1", "hash"));
-        assert!(!state.message_processed("turn-2", "hash"));
-    }
-
-    #[test]
-    fn internal_turns_cannot_revoke_a_user_edit_constraint() {
-        let protected = constraint("don't touch tests", ConstraintMatcher::TestFiles);
-        let revocation = ConstraintRevocation {
-            constraint_id: protected.id.clone(),
-            description: "tests may be modified now".to_string(),
-        };
-
-        let (revoked, unmatched) =
-            validated_revocation_ids(&[revocation], &[protected.clone()], false);
-
-        assert!(revoked.is_empty());
-        assert!(unmatched.is_empty());
-        let (revoked, unmatched) = validated_revocation_ids(
-            &[ConstraintRevocation {
-                constraint_id: protected.id.clone(),
-                description: "tests may be modified now".to_string(),
-            }],
-            &[protected],
-            true,
-        );
-        assert_eq!(revoked, vec!["test:don't touch tests".to_string()]);
-        assert!(unmatched.is_empty());
-    }
-
-    #[test]
-    fn state_applies_only_validated_explicit_revocations() {
-        let protected = constraint("don't touch tests", ConstraintMatcher::TestFiles);
-        let protected_id = protected.id.clone();
-        let mut state = EditConstraintState::default();
-        state.constraints.push(protected);
-
-        state.merge_extraction(ConstraintExtractionRecord {
-            message_sha256: "relaxation-hash".to_string(),
-            dialog_turn_id: Some("turn-2".to_string()),
-            status: ExtractionStatus::Extracted,
-            constraints: Vec::new(),
-            deterministic_constraint_count: 0,
-            model_attempts: 1,
-            active_constraint_ids: vec![protected_id.clone()],
-            revocation_authorized: true,
-            model_status: ModelExtractionStatus::Parsed,
-            model_constraints: Vec::new(),
-            model_revocations: vec![ConstraintRevocation {
-                constraint_id: protected_id.clone(),
-                description: "tests may be modified now".to_string(),
-            }],
-            revoked_constraint_ids: vec![protected_id],
-            unmatched_revocation_ids: Vec::new(),
-            input_chars: 24,
-            prompt_chars: 24,
-            input_truncated: false,
-            latency_ms: 1,
-            extracted_at_ms: 1,
-            failure: None,
-            response_excerpt: Some(
-                r#"{"additions":[],"revocations":[{"constraint_id":"test:don't touch tests"}]}"#
-                    .to_string(),
-            ),
-        });
-
-        assert!(state.constraints.is_empty());
-        assert_eq!(state.schema_version, EDIT_CONSTRAINT_SCHEMA_VERSION);
-    }
-
-    #[test]
-    fn failed_or_unmatched_revocation_keeps_active_constraint() {
-        let protected = constraint("don't touch tests", ConstraintMatcher::TestFiles);
-        let mut state = EditConstraintState::default();
-        state.constraints.push(protected.clone());
-
-        state.merge_extraction(ConstraintExtractionRecord {
-            message_sha256: "invalid-relaxation-hash".to_string(),
-            dialog_turn_id: Some("turn-2".to_string()),
-            status: ExtractionStatus::NoConstraints,
-            constraints: Vec::new(),
-            deterministic_constraint_count: 0,
-            model_attempts: 1,
-            active_constraint_ids: vec![protected.id.clone()],
-            revocation_authorized: true,
-            model_status: ModelExtractionStatus::Parsed,
-            model_constraints: Vec::new(),
-            model_revocations: vec![ConstraintRevocation {
-                constraint_id: "invented-id".to_string(),
-                description: "ambiguous relaxation".to_string(),
-            }],
-            revoked_constraint_ids: Vec::new(),
-            unmatched_revocation_ids: vec!["invented-id".to_string()],
-            input_chars: 20,
-            prompt_chars: 20,
-            input_truncated: false,
-            latency_ms: 1,
-            extracted_at_ms: 1,
-            failure: None,
-            response_excerpt: None,
-        });
-
-        assert_eq!(state.constraints, vec![protected]);
-    }
-
-    #[test]
-    fn find_violation_returns_first_match() {
-        let constraints = vec![
-            constraint("don't touch tests", ConstraintMatcher::TestFiles),
-            constraint(
-                "don't touch lockfiles",
-                ConstraintMatcher::Extension {
-                    exts: vec![".lock".to_string()],
-                },
-            ),
-        ];
-        assert_eq!(
-            find_violation(&constraints, "report/util_test.go")
-                .map(|constraint| constraint.description.as_str()),
-            Some("don't touch tests")
-        );
-        assert_eq!(
-            find_violation(&constraints, "Cargo.lock")
-                .map(|constraint| constraint.description.as_str()),
-            Some("don't touch lockfiles")
-        );
-    }
-
-    #[test]
-    fn new_files_are_exempt_only_from_test_file_constraints() {
-        let test_only = EditConstraintState {
-            constraints: vec![constraint(
-                "don't touch tests",
-                ConstraintMatcher::TestFiles,
-            )],
-            ..Default::default()
-        };
-        let test_path = vec!["test/repro-test.ts".to_string()];
-        assert!(has_only_relaxable_test_file_violations(
-            &test_only, &test_path, "write"
-        ));
-
-        let stricter = EditConstraintState {
-            constraints: vec![
-                constraint("don't touch tests", ConstraintMatcher::TestFiles),
-                constraint(
-                    "don't modify generated files",
-                    ConstraintMatcher::PathUnderDir {
-                        dirs: vec!["test".to_string()],
-                    },
-                ),
-            ],
-            ..Default::default()
-        };
-        assert!(!has_only_relaxable_test_file_violations(
-            &stricter, &test_path, "write"
-        ));
-    }
-
-    #[test]
-    fn agent_created_test_files_can_be_cleaned_up_but_delete_only_rules_remain_strict() {
-        let path = vec!["test/repro-test.ts".to_string()];
-        let mut state = EditConstraintState {
-            constraints: vec![constraint(
-                "don't modify tests",
-                ConstraintMatcher::TestFiles,
-            )],
-            ..Default::default()
-        };
-        state.remember_agent_created_paths(path.clone(), "turn-1");
-        assert!(can_mutate_agent_created_test_file(
-            Some(&state),
-            &path,
-            "edit",
-            false
-        ));
-        assert!(can_mutate_agent_created_test_file(
-            Some(&state),
-            &path,
-            "delete",
-            false
-        ));
-
-        state.constraints.push(ExtractedConstraint {
-            id: "test:do-not-delete".to_string(),
-            description: "don't delete tests".to_string(),
-            operation_scope: ConstraintOperationScope::DeleteOnly,
-            matcher: ConstraintMatcher::TestFiles,
-            source: ConstraintSource::Deterministic,
-            source_text: Some("Do not delete tests.".to_string()),
-        });
-        assert!(
-            find_violation_for_operation(&state.constraints, "test/repro-test.ts", "delete")
-                .is_some()
-        );
-        assert!(
-            find_violation_for_operation(&state.constraints, "test/repro-test.ts", "edit")
-                .is_some()
-        );
-        assert!(!can_mutate_agent_created_test_file(
-            Some(&state),
-            &path,
-            "delete",
-            false
-        ));
-
-        state.forget_agent_created_paths_under(&path);
-        assert!(!state.is_agent_created_path(&path));
-    }
-
-    #[test]
-    fn rollback_discards_future_constraints_and_helper_provenance() {
-        let initial_constraint = constraint("don't modify tests", ConstraintMatcher::TestFiles);
-        let future_constraint = constraint(
-            "don't modify lockfiles",
-            ConstraintMatcher::Extension {
-                exts: vec![".lock".to_string()],
-            },
-        );
-        let mut state = EditConstraintState::default();
-        state.merge_extraction(ConstraintExtractionRecord {
-            message_sha256: "turn-1-hash".to_string(),
-            dialog_turn_id: Some("turn-1".to_string()),
-            status: ExtractionStatus::Extracted,
-            constraints: vec![initial_constraint.clone()],
-            deterministic_constraint_count: 1,
-            model_attempts: 0,
-            active_constraint_ids: Vec::new(),
-            revocation_authorized: true,
-            model_status: ModelExtractionStatus::NotRun,
-            model_constraints: Vec::new(),
-            model_revocations: Vec::new(),
-            revoked_constraint_ids: Vec::new(),
-            unmatched_revocation_ids: Vec::new(),
-            input_chars: 10,
-            prompt_chars: 10,
-            input_truncated: false,
-            latency_ms: 1,
-            extracted_at_ms: 1,
-            failure: None,
-            response_excerpt: None,
-        });
-        state.remember_agent_created_paths(vec!["tests/kept_repro.rs".to_string()], "turn-1");
-        state.merge_extraction(ConstraintExtractionRecord {
-            message_sha256: "turn-2-hash".to_string(),
-            dialog_turn_id: Some("turn-2".to_string()),
-            status: ExtractionStatus::Extracted,
-            constraints: vec![future_constraint],
-            deterministic_constraint_count: 0,
-            model_attempts: 1,
-            active_constraint_ids: vec![initial_constraint.id.clone()],
-            revocation_authorized: true,
-            model_status: ModelExtractionStatus::Parsed,
-            model_constraints: Vec::new(),
-            model_revocations: Vec::new(),
-            revoked_constraint_ids: vec![initial_constraint.id.clone()],
-            unmatched_revocation_ids: Vec::new(),
-            input_chars: 10,
-            prompt_chars: 10,
-            input_truncated: false,
-            latency_ms: 1,
-            extracted_at_ms: 2,
-            failure: None,
-            response_excerpt: None,
-        });
-        state.remember_agent_created_paths(vec!["tests/future_repro.rs".to_string()], "turn-2");
-
-        state.rollback_to_surviving_turns(&HashSet::from(["turn-1".to_string()]));
-
-        assert_eq!(state.constraints, vec![initial_constraint]);
-        assert_eq!(state.extractions.len(), 1);
-        assert_eq!(
-            state.agent_created_paths,
-            vec!["tests/kept_repro.rs".to_string()]
-        );
-        assert_eq!(state.agent_created_path_records.len(), 1);
-        assert_eq!(state.agent_created_path_records[0].dialog_turn_id, "turn-1");
-    }
-
-    #[test]
-    fn deterministic_extractor_marks_explicit_test_deletion_as_delete_only() {
-        let constraint = deterministic_test_constraint("Do not delete test files.")
-            .expect("explicit test deletion should be extracted");
-        assert_eq!(
-            constraint.operation_scope,
-            ConstraintOperationScope::DeleteOnly
-        );
-        assert!(constraint.matcher.matches("tests/example.rs"));
-    }
-
-    #[test]
-    fn force_is_rejected_even_without_runtime_context() {
-        let rejection =
-            check(None, "Edit", "edit", "tests/example.rs", true).expect("force must be denied");
-        assert!(!rejection.result);
-        assert_eq!(rejection.error_code, Some(403));
-        assert_eq!(
-            rejection
-                .meta
-                .as_ref()
-                .and_then(|value| value.get("guard_decision"))
-                .and_then(Value::as_str),
-            Some("force_denied")
-        );
-    }
-
-    #[tokio::test]
-    async fn blank_input_is_no_constraints_not_failure() {
-        let extraction = extract_constraints("   \n  ").await;
-        assert_eq!(extraction.status, ExtractionStatus::NoConstraints);
-        assert!(extraction.constraints.is_empty());
-        assert!(extraction.failure.is_none());
-    }
-
-    #[test]
-    fn successful_mutation_telemetry_is_persisted_as_jsonl() {
-        let root = std::env::temp_dir().join(format!(
-            "bitfun-edit-constraint-telemetry-{}",
-            Uuid::new_v4()
-        ));
-        fs::create_dir_all(&root).expect("create temp workspace");
-        let event = json!({
-            "event": "mutation_applied",
-            "tool_call_id": "tool-call-1",
-            "requested_path": "tests/example.rs",
-        });
-        let telemetry_path = root.join(TELEMETRY_RELATIVE_PATH);
-        append_jsonl(&telemetry_path, &event).expect("append telemetry event");
-        let line = fs::read_to_string(&telemetry_path).expect("read telemetry");
-        let event: Value = serde_json::from_str(line.trim()).expect("valid jsonl event");
-        assert_eq!(event["event"], "mutation_applied");
-        assert_eq!(event["tool_call_id"], "tool-call-1");
-        assert_eq!(event["requested_path"], "tests/example.rs");
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn local_recursive_delete_fallback_finds_protected_descendant() {
-        let root = std::env::temp_dir().join(format!(
-            "bitfun-edit-constraint-recursive-delete-{}",
-            Uuid::new_v4()
-        ));
-        let target = root.join("parent");
-        fs::create_dir_all(target.join("tests")).expect("create test directory");
-        fs::write(target.join("tests/example.rs"), "test").expect("create test file");
-        let context = ToolUseContext {
-            tool_call_id: Some("tool-call-1".to_string()),
-            agent_type: Some("agentic".to_string()),
-            session_id: None,
-            dialog_turn_id: Some("turn-1".to_string()),
-            workspace: Some(WorkspaceBinding::new(None, root.clone())),
-            unlocked_collapsed_tools: Vec::new(),
-            custom_data: HashMap::new(),
-            computer_use_host: None,
-            runtime_tool_restrictions: ToolRuntimeRestrictions::default(),
-            runtime_handles: bitfun_runtime_ports::ToolRuntimeHandles::default(),
-        };
-        let state = EditConstraintState {
-            constraints: vec![constraint(
-                "don't touch tests",
-                ConstraintMatcher::TestFiles,
-            )],
-            ..Default::default()
-        };
-
-        let rejection =
-            check_local_recursive_delete(&context, "parent", &target.to_string_lossy(), &state)
-                .expect("recursive delete should be denied");
-        assert_eq!(rejection.error_code, Some(403));
-        assert!(rejection
-            .message
-            .as_deref()
-            .unwrap_or_default()
-            .contains("tests"));
-
-        let _ = fs::remove_dir_all(root);
-    }
-}
+mod tests;
